backend-python/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python
imagehash
torch
torchvision
pydantic